easy/0874WalkingRobotSimulation/robot_sim_test.go
package robotsim

import "testing"

func Test_robotSim(t *testing.T) {
	type args struct {
		commands  []int
		obstacles [][]int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// TODO: Add test cases.
		{
			name: "test1",
			args: args{
				commands: []int{4, -1, 4, -2, 4},
				obstacles: [][]int{
					{2, 4},
				},
			},
			want: 65,
		},
		{
			name: "test2",
			args: args{
				commands: []int{1, 2, -2, 5, -1, -2, -1, 8, 3, -1, 9, 4, -2, 3, 2, 4, 3, 9, 2, -1, -1, -2, 1, 3, -2, 4, 1, 4, -1, 1, 9, -1, -2, 5, -1, 5, 5, -2, 6, 6},
				obstacles: [][]int{
					{-57, -58},
					{-72, 91},
				},
			},
			want: 2845,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := robotSim(tt.args.commands, tt.args.obstacles); got != tt.want {
				t.Errorf("robotSim() = %v, want %v", got, tt.want)
			}
		})
	}
}
